--- src/main/java/com/playtomic/tests/wallet/service/StripeAmountTooSmallException.java
@@ -1,4 +1,0 @@
-package com.playtomic.tests.wallet.service;
-
-public class StripeAmountTooSmallException extends StripeServiceException {
-}
