--- src/main/java/com/playtomic/tests/wallet/service/StripeServiceException.java
@@ -1,4 +1,0 @@
-package com.playtomic.tests.wallet.service;
-
-public class StripeServiceException extends RuntimeException {
-}
