+++ src/main/java/com/playtomic/tests/wallet/service/payment/stripe/StripeAmountTooSmallException.java
@@ -1,0 +1,10 @@
+package com.playtomic.tests.wallet.service.payment.stripe;
+
+/**
+ * Updated by Orkun Cavdar on 08/02/2022
+ */
+public class StripeAmountTooSmallException extends StripeServiceException {
+    public StripeAmountTooSmallException(String message) {
+        super(message);
+    }
+}
