+++ src/main/java/com/playtomic/tests/wallet/service/payment/stripe/StripeRestTemplateResponseErrorHandler.java
@@ -1,4 +1,4 @@
-package com.playtomic.tests.wallet.service;
+package com.playtomic.tests.wallet.service.payment.stripe;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.client.ClientHttpResponse;
@@ -6,12 +6,15 @@ import org.springframework.web.client.DefaultResponseErrorHandler;
 
 import java.io.IOException;
 
+/**
+ * Updated by Orkun Cavdar on 08/02/2022
+ */
 public class StripeRestTemplateResponseErrorHandler extends DefaultResponseErrorHandler {
 
     @Override
     protected void handleError(ClientHttpResponse response, HttpStatus statusCode) throws IOException {
         if (statusCode == HttpStatus.UNPROCESSABLE_ENTITY) {
-            throw new StripeAmountTooSmallException();
+            throw new StripeAmountTooSmallException("Amount is too small to charge");
         }
 
         super.handleError(response, statusCode);
