+++ src/main/java/com/playtomic/tests/wallet/service/payment/stripe/StripeService.java
@@ -1,9 +1,12 @@
-package com.playtomic.tests.wallet.service;
+package com.playtomic.tests.wallet.service.payment.stripe;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.playtomic.tests.wallet.service.payment.PaymentService;
 import com.sun.istack.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.NonNull;
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.stereotype.Service;
@@ -15,12 +18,15 @@ import java.net.URI;
 
 /**
  * Handles the communication with Stripe.
- *
+ * <p>
  * A real implementation would call to String using their API/SDK.
  * This dummy implementation throws an error when trying to charge less than 10€.
+ * Updated by Orkun Cavdar on 08/02/2022
  */
 @Service
-public class StripeService {
+public class StripeService implements PaymentService {
+
+    private static final Log log = LogFactory.getLog(StripeService.class);
 
     @NonNull
     private URI chargesUri;
@@ -31,8 +37,8 @@ public class StripeService {
     @NonNull
     private RestTemplate restTemplate;
 
-    public StripeService(@Value("stripe.simulator.charges-uri") @NonNull URI chargesUri,
-                         @Value("stripe.simulator.refunds-uri") @NonNull URI refundsUri,
+    public StripeService(@Value("${stripe.simulator.charges-uri}") @NonNull URI chargesUri,
+                         @Value("${stripe.simulator.refunds-uri}") @NonNull URI refundsUri,
                          @NotNull RestTemplateBuilder restTemplateBuilder) {
         this.chargesUri = chargesUri;
         this.refundsUri = refundsUri;
@@ -44,7 +50,7 @@ public class StripeService {
 
     /**
      * Charges money in the credit card.
-     *
+     * <p>
      * Ignore the fact that no CVC or expiration date are provided.
      *
      * @param creditCardNumber The number of the credit card
