+++ src/main/java/com/playtomic/tests/wallet/service/payment/stripe/StripeServiceException.java
@@ -1,0 +1,10 @@
+package com.playtomic.tests.wallet.service.payment.stripe;
+
+/**
+ * Updated by Orkun Cavdar on 08/02/2022
+ */
+public class StripeServiceException extends RuntimeException {
+    public StripeServiceException(String message) {
+        super(message);
+    }
+}
